move decode_user to utils.py
