+++ backend/ops_api/ops/auth/auth.py
@@ -1,36 +1,13 @@
 from functools import wraps
 from typing import Callable, Optional
 
-from authlib.jose import JsonWebToken
 from flask import Response
 from flask_jwt_extended import get_current_user, get_jwt_identity, jwt_required
 
 from ops_api.ops.auth.auth_enum import Permission, PermissionType
 from ops_api.ops.auth.authorization import AuthorizationGateway, BasicAuthorizationPrivider
-from ops_api.ops.auth.utils import get_jwks
 from ops_api.ops.utils.errors import error_simulator
 from ops_api.ops.utils.response import make_response_with_headers
-
-
-def decode_user(
-    payload: Optional[str] = None,
-    provider: Optional[str] = None,
-) -> dict[str, str]:
-    # TODO: Determine which claims we want to validate when decoding a user from the provider
-    #       beyond just the signature. Should these be universal for any claims (global)?
-    # ex: validate the JTI, validate the expiration, validate the si
-    # claims_options = {
-    #     "iss": {
-    #         "essential": True,
-    #         "values": current_app.config["AUTHLIB_OAUTH_CLIENTS"][provider]["client_id"],
-    #     },
-    #     "jti": {"validate": JWTClaims.validate_jti},
-    #     "exp": {"validate": JWTClaims.validate_exp},
-    # }
-    jwt = JsonWebToken(["RS256"])
-    # claims = jwt.decode(payload, get_jwks(provider), claims_options=claims_options)
-    claims = jwt.decode(payload, get_jwks(provider))
-    return claims
 
 
 class ExtraCheckError(Exception):
