+++ backend/ops_api/ops/auth/exceptions.py
@@ -1,0 +1,6 @@
+class NotActiveUserError(Exception):
+    """
+    Exception to raise when the user is not active.
+    """
+
+    pass
